requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email-validator
WTForms
pandas
openpyxl

Pillow>=10.3.0

requests
beautifulsoup4
